talentscout-chatbot/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
regex>=2023.0.0
cryptography>=41.0.0
